Fix 64-bit bug in UtlIntPtr - trying to exchange two 64-bit values using 32-bit temp variable. Also some minor warning fixes.
git-svn-id: http://scm.sipfoundry.org/rep/sipX/main@10958 a612230a-c5fa-0310-af8b-88eea846685b

+++ sipXportLib/src/utl/UtlIntPtr.cpp
@@ -90,7 +90,7 @@ UtlIntPtr& UtlIntPtr::operator=(const UtlIntPtr& rhs)
 
 intptr_t UtlIntPtr::setValue(intptr_t iValue)
 {
-    int iOldValue = mValue ;
+    intptr_t iOldValue = mValue ;
     mValue = iValue ;
 
     return iOldValue ;
@@ -106,7 +106,7 @@ intptr_t UtlIntPtr::getValue() const
 
 unsigned UtlIntPtr::hash() const
 {
-   return mValue ; 
+   return (unsigned)mValue ; 
 }
 
 
@@ -125,7 +125,7 @@ int UtlIntPtr::compareTo(UtlContainable const * inVal) const
     {
         UtlIntPtr* temp = (UtlIntPtr*)inVal ; 
         intptr_t inInt = temp->getValue() ; 
-        result = mValue - inInt ; 
+        result = (int)(mValue - inInt); 
     }
     else
     {
